requirements.txt
python-telegram-bot[job-queue]>=21.0
python-dotenv
aiosqlite>=0.19.0
aiohttp>=3.9.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
psycopg[binary]>=3.1.18
psycopg_pool>=3.2.0
